requirement.txt
bcrypt==4.0.1
blinker==1.6.2
click==8.1.4
colorama==0.4.6
dnspython==2.3.0
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-PyMongo==2.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.25.1
opencv-python==4.8.0.74
PyJWT==2.7.0
pymongo==4.4.1
python-dotenv==1.0.0
Werkzeug==2.3.6
